Revert to use WebCredentials as password storage.
