+++ UnstoppableWallet/UnstoppableWallet/Core/Crypto/BackupCrypto.swift
@@ -129,3 +129,12 @@ extension BackupCrypto {
         case simplePassword
     }
 }
+
+extension BackupCrypto.ValidationError: LocalizedError {
+    public var errorDescription: String? {
+        switch self {
+        case .emptyPassphrase: return "backup.cloud.password.error.empty_passphrase".localized
+        case .simplePassword: return "backup.cloud.password.error.minimum_requirement".localized
+        }
+    }
+}
