+++ UnstoppableWallet/UnstoppableWallet/Modules/Backup/Password/Storage/BackupKeychainService.swift
@@ -3,6 +3,8 @@ import Foundation
 class BackupKeychainService: IBackupPasswordStorage {
     private static let service = "unstoppable.money"
 
+    var requiresUnlock: Bool { true }
+
     func save(password: String, account: String) async throws {
         guard let data = password.data(using: .utf8) else {
             throw StorageError.encodingFailed
